+++ src/api/auth.ts
@@ -1,13 +1,12 @@
+import { error } from "console";
 import instance from "./axios";
-import { AxiosError } from "axios";
+
 
 interface PostSignUpQuery {
-  body: {
     email: string;
-    name: string;
+    name?: string;
     password: string;
-    passwordConfirmation: string;
-  };
+    passwordConfirmation?: string;
 }
 
 interface PostSignInQuery {
@@ -23,24 +22,19 @@ interface PostRefreshTokenQuery {
 
 // 회원가입
 export const postSignUp = async (body: PostSignUpQuery) => {
-  try {
     const res = await instance.post(`/auth/signUp`, body);
     if (res.status === 200) {
       return res.data;
-    }
-  } catch (err) {
-    console.error("회원가입에 실패했습니다.", err);
-    return {};
+    } 
   }
-};
 
 // 로그인
 export const postSignIn = async (body: PostSignInQuery) => {
     const res = await instance.post(`/auth/signIn`, body);
     if (res.status === 200) {
       return res.data;
-    
 };
+}
 
 // 토큰 재갱신 
 export const postRefreshToken = async (body: PostRefreshTokenQuery) => {
