+++ src/containers/LoginForm.tsx
@@ -1,72 +1,53 @@
 import { useForm } from "react-hook-form";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import instance from "@/api/axios";
-import { InputVlaues } from "./LoginForm";
-import Button from "./ui/Button";
+import { useState } from "react";
+import useAuthStore from "@/store/AuthStore";
+import { AxiosError } from "axios";
+import FilledButton from "@/components/ui/Button/FilledButton";
 
-const SignUpForm = () => {
+export interface InputVlaues {
+  email: string;
+  name?: string ;
+  password: string;
+  passwordConfirmation?: string;
+}
+
+const LoginForm = () => {
   const {
     register,
     handleSubmit,
-    watch,
     formState: { isSubmitting, errors },
   } = useForm<InputVlaues>({
     mode: "onSubmit",
   });
-
-  const password = watch("password");
+  const [submitError, setSubmitError] = useState("");
   const router = useRouter();
+  const { login } = useAuthStore();
 
   const onSubmit = async (data: InputVlaues) => {
-    await instance.post("/auth/signUp", {
-      email: data.email,
-      name: data.name,
-      password: data.password,
-      passwordConfirmation: data.password,
-    });
-
-    window.alert("가입이 완료되었습니다"); 
-
-    router.push("/login");
+    try {
+      await login(data.email, data.password);
+      router.push("/");
+    } catch (error) {
+      if (error instanceof AxiosError) {
+        setSubmitError(error.response?.data.message);
+      }
+    }
   };
 
+  const handleChange = () => {
+    setSubmitError("");
+  };
   return (
     <div className="flex items-center justify-center min-h-screen Mobile:px-5 bg-gray-50">
       <div className="w-full max-w-md">
-        <h2 className="text-2xl font-semibold mb-[50px] text-center">
-          회원가입
-        </h2>
+        <h2 className="text-2xl font-semibold mb-[50px] text-center">로그인</h2>
         <form
           noValidate
           onSubmit={handleSubmit(onSubmit)}
           className="flex flex-col gap-6 mb-10"
         >
-          <div className="flex flex-col gap-2.5">
-            <label htmlFor="name" className="text-md text-gray-500">
-              이름
-            </label>
-            <input
-              className={`rounded-[10px] px-5 py-[10.5px] mb-[10px]gray100 bg-gray-100 placeholder:text-md placeholder:text-gray-400 ${
-                errors.email ? "outline-red-200" : "outline-green-200"
-              }`}
-              id="name"
-              type="text"
-              placeholder="이름을 입력해 주세요"
-              {...register("name", {
-                required: true,
-                maxLength: {
-                  value: 10,
-                  message: "열 자 이하로 작성해 주세요",
-                },
-              })}
-            ></input>
-            {errors.name && (
-              <span className="text-xs text-red-200" role="alert">
-                {errors.name.message}
-              </span>
-            )}
-          </div>
           <div className="flex flex-col gap-2.5">
             <label htmlFor="email" className="text-md text-gray-500">
               이메일
@@ -85,6 +66,7 @@ const SignUpForm = () => {
                   message: "이메일 형식으로 작성해 주세요.",
                 },
               })}
+              onChange={handleChange}
             ></input>
             {errors.email && (
               <span className="text-xs text-red-200" role="alert">
@@ -113,51 +95,30 @@ const SignUpForm = () => {
                   message: "8자 이상 작성해 주세요",
                 },
               })}
+              onChange={handleChange}
             ></input>
             {errors.password && (
               <span className="text-xs text-red-200" role="alert">
                 {errors.password.message}
               </span>
             )}
-          </div>
-          <div className="flex flex-col gap-2.5">
-            <label
-              htmlFor="passwordConfirmation"
-              className="text-md font-normal text-gray-500"
-            >
-              비밀번호 확인
-            </label>
-            <input
-              className={`rounded-[10px] px-5 py-[10.5px] mb-[10px] bg-gray-100 placeholder:text-md placeholder:text-gray4-00 ${
-                errors.password ? "outline-red-200" : "outline-green-200"
-              }`}
-              id="passwordConfirmation"
-              type="password"
-              placeholder="비밀번호를 입력해주세요"
-              {...register("passwordConfirmation", {
-                required: true,
-                validate: (value) =>
-                  value === password || "비밀번호가 일치하지 않습니다",
-              })}
-            ></input>
-            {errors.passwordConfirmation && (
+            {submitError && (
               <span className="text-xs text-red-200" role="alert">
-                {errors.passwordConfirmation.message}
+                {submitError}
               </span>
             )}
           </div>
-          <Button
+          <FilledButton
             type="submit"
             disabled={isSubmitting}
             onClick={handleSubmit(onSubmit)}
           >
-            회원가입
-          </Button>
-        </form>
-        <div className="flex justify-center gap-[10px]">
-          <span className="text-md text-gray-400">이미 회원이신가요?</span>
-          <Link href="/login" className="text-md font-normal text-green-200">
             로그인
+          </FilledButton>
+        </form>
+        <div className="flex justify-center">
+          <Link href="/login" className="text-md font-normal text-green-200">
+            회원가입
           </Link>
         </div>
       </div>
@@ -165,4 +126,4 @@ const SignUpForm = () => {
   );
 };
 
-export default SignUpForm;
+export default LoginForm;
