+++ src/pages/_app.tsx
@@ -5,7 +5,7 @@ import type { AppProps } from "next/app";
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <>
-      <Header />
+      {/* <Header /> */}
       <Component {...pageProps} />
     </>
   );
